Updating pytest and pydantic tests
