+++ python_pydantic_test/Readme.md
@@ -1,0 +1,10 @@
+<h1>Description</h1>
+This folder considers the different implementation of testing pydantic library. </br>
+Checked with pep8 and pycodestyle.</br>
+</br>
+</br>
+The pydantic library checking is implemented within file python_pydantic_for_api.py</br>
+Run example: python python_pydantic_for_api.py</br>
+Remark: pep8 checking fails E701 multiple statements on one line (colon) but passes for pycodestyle</br>
+</br>
+</br>
